358. Duplicating Cabins

+++ src/services/apiCabins.js
@@ -11,12 +11,26 @@ export async function apiGetCabins() {
   return data;
 }
 
-export async function apiCreateCabin(newCabin) {
+export async function apiCreateEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll('/', '');
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  // 1. Создание cabin
-  let query = supabase.from('cabins').insert([{ ...newCabin, image: imagePath }]);
+  // 1. Создание/редактирование cabin
+  let query = supabase.from('cabins');
+
+  // a) Создание
+  if (!id) {
+    query = query.insert([{ ...newCabin, image: imagePath }]);
+  }
+
+  // b) Редактирование
+  if (id) {
+    query = query.update({ ...newCabin, image: imagePath }).eq('id', id);
+  }
+
   const { data, error } = await query.select().single();
 
   if (error) {
@@ -25,11 +39,14 @@ export async function apiCreateCabin(newCabin) {
   }
 
   // 2. Загрузка картинки
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from('cabin-images')
     .upload(imageName, newCabin.image);
 
-  if (storageError) {
+  // 3. Удаление cabin если была ошибка загрузки картинки
+  if (storageError && !id) {
     await supabase.from('cabins').delete().eq('id', data.id);
     console.error(storageError);
     throw new Error('Cabin image could not be uploaded and the cabin was not created');
@@ -38,46 +55,8 @@ export async function apiCreateCabin(newCabin) {
   return data;
 }
 
-export async function apiEditCabin(newCabin, id, editImage) {
-  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
-  let query = supabase.from('cabins');
-  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll('/', '');
-  const imagePath = hasImagePath
-    ? newCabin.image
-    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
-
-  // Если hasImagePath === true то это редактирование текста
-  // Если hasImagePath === falsе то это редактирование текста и загрузка новой картинки
-  if (hasImagePath) {
-    query = query.update(newCabin).eq('id', id);
-  } else {
-    query = query.update({ ...newCabin, image: imagePath }).eq('id', id);
-  }
-
-  const { data, error } = await query.select().single();
-
-  if (error) {
-    console.error(error);
-    throw new Error('Cabin could not be edited');
-  }
-
-  if (!hasImagePath) {
-    await supabase.storage.from('cabin-images').remove([editImage]);
-    const { error: storageError } = await supabase.storage
-      .from('cabin-images')
-      .upload(imageName, newCabin.image);
-    if (storageError) {
-      console.error(storageError);
-      throw new Error('Cabin image could not be uploaded');
-    }
-  }
-
-  return data;
-}
-
-export async function apiDeleteCabin(id, image) {
+export async function apiDeleteCabin(id) {
   const { error } = await supabase.from('cabins').delete().eq('id', id);
-  await supabase.storage.from('cabin-images').remove([image]);
 
   if (error) {
     console.error(error);
